requirements.txt
numpy
Pillow
ultralytics
transformers>=4.57.0
huggingface_hub
torch
accelerate
timm
